+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- **Table** with FilterBar example.
+- **FilterBar** examples that were broken.
+
+### Added
+
+- **Conditions** component now it's not EXPERIMENTAL anymore.
+- **Statement** refactor so it can be shared between FilterTag and Conditions components.
+
 ## [9.77.5] - 2019-09-05
 
 ### Fixed
@@ -14,6 +24,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `Input` disabled background.
 
 ## [9.77.4] - 2019-09-04
+
 ### Changed
 - Move `appearance: none` inline style to class in `Input` component.